refactored endpointrequest objects finished up streamclient
