+++ oanda_dotnet/model/StreamEndpointRequest.cs
@@ -6,5 +6,10 @@ namespace oanda_dotnet.model
     {
         public abstract string Endpoint { get; }
         public abstract Method Method { get; }
+
+        public string Resource()
+        {
+            return string.Empty;
+        }
     }
 }
